[impl] Package client-py: src-layout, console script, version from CHANGES.md
- src-layout client-py/src/arduino_esp32_tft_terminal/ (config.py, app/, lib/,
  run.py->cli.py); flat config/app/lib imports rewritten to the package namespace
- cli.py: wrap module-level code into main(); add __main__ guard and __main__.py
- pyproject [project]: hatchling, dynamic version from CHANGES.md regex,
  deps numpy/pyserial/claude-busy-monitor, console script arduino-esp32-tft-terminal
- move CHANGES.md into client-py/ (hatchling version source must sit by pyproject)
- update READMEs + CLAUDE.md for the new invocation; add claude-busy-monitor to requirements.txt

Verified: uv build, wheel install in a venv, console-script + python -m --help (exit 0),
version 0.1.0, ruff green.

Closes #17

Authored-By: Claude Opus 4.8 (1M context) <noreply@anthropic.com>
