+++ client-py/requirements.txt
@@ -1,3 +1,3 @@
 numpy>=1.26.1
 pyserial>=3.5
-
+claude-busy-monitor
