+++ client-py/src/arduino_esp32_tft_terminal/__init__.py
@@ -1,0 +1,1 @@
+"""Arduino ESP32 TFT Terminal — Python client."""
