+++ client-py/src/arduino_esp32_tft_terminal/__main__.py
@@ -1,0 +1,3 @@
+from arduino_esp32_tft_terminal.cli import main
+
+main()
