+++ client-py/src/arduino_esp32_tft_terminal/app/collisions.py
@@ -14,10 +14,10 @@ from typing import Callable
 
 import numpy as np
 
-import config
-from app import App, TimeEscaper
-from lib.board import Board
-from lib.gfx import Gfx
+from arduino_esp32_tft_terminal import config
+from arduino_esp32_tft_terminal.app import App, TimeEscaper
+from arduino_esp32_tft_terminal.lib.board import Board
+from arduino_esp32_tft_terminal.lib.gfx import Gfx
 
 COLOR_BORDER = 127, 127, 125
 Color = tuple[int, int, int]
