+++ client-py/src/arduino_esp32_tft_terminal/app/monitor_host.py
@@ -1,9 +1,9 @@
 import socket
 import subprocess
 
-import config
-from app import App, TimeEscaper
-from lib.board import Board
+from arduino_esp32_tft_terminal import config
+from arduino_esp32_tft_terminal.app import App, TimeEscaper
+from arduino_esp32_tft_terminal.lib.board import Board
 
 CHOICE_EXIT = 1
 CHOICE_NEXT = 2
