+++ client-py/src/arduino_esp32_tft_terminal/app/tunnel.py
@@ -1,8 +1,8 @@
 import math
 
-import config
-from app import App, TimeEscaper
-from lib.board import Board
+from arduino_esp32_tft_terminal import config
+from arduino_esp32_tft_terminal.app import App, TimeEscaper
+from arduino_esp32_tft_terminal.lib.board import Board
 
 K = 0.65
 NB = 12
